requirements.txt
Django==1.9.4
django-cors-headers==1.1.0
django-environ==0.4.0
django-filter==1.1.0
djangorestframework==3.6.4
psycopg2==2.6.1
requests==2.9.1
python-memcached==1.59
